Force read after write operation on socket
No read is performed after a write. This could be a problem
if the instrument returns an acknowledge string and is not
removed from the buffer.

Force a read (acknowledge) after a write in the
bySocket class.

Need to subclass Communicator from object to have access
to super method in the new notation.